requirements.txt
# Core dependencies
Flask==2.0.1
flask-cors==3.0.1
gunicorn==20.1.0

# AI/ML dependencies
torch==2.0.1
transformers==4.30.2
sentencepiece==0.1.99
scikit-learn==1.0.2
numpy==1.24.3

# News API and requests
newsapi-python==0.2.7
requests==2.31.0
beautifulsoup4==4.12.2

# Text processing
nltk==3.8.1
regex==2023.5.5

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
pandas==2.0.2
urllib3==2.0.3
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
six==1.16.0
packaging==23.1

# Optional but recommended
pytest==7.3.1  # for testing
black==23.3.0  # for code formatting
pylint==2.17.4  # for code linting 